+++ src/basic/components/pages/CartPage.tsx
@@ -1,20 +1,25 @@
+import { useCallback } from 'react';
+
 import { CartItem, Coupon } from '../../../types';
 import { ProductWithUI } from '../../App';
+import CartItemList from '../cart/CartItemList';
+import CouponSelector from '../cart/CouponSelector';
+import EmptyCart from '../cart/EmptyCart';
+import ProductList from '../cart/ProductList';
+type NotificationType = 'error' | 'success' | 'warning';
 
 type CartPageProps = {
   products: ProductWithUI[];
   filteredProducts: ProductWithUI[];
-  debouncedSearchTerm: string;
+  searchProductName: string;
 
   cart: CartItem[];
-  addToCart: (product: ProductWithUI) => void;
-  removeFromCart: (productId: string) => void;
-  updateQuantity: (productId: string, quantity: number) => void;
+
   calculateItemTotal: (item: CartItem) => number;
   getRemainingStock: (product: ProductWithUI) => number;
 
   formatPrice: (price: number, productId?: string) => string;
-
+  setCart: React.Dispatch<React.SetStateAction<CartItem[]>>;
   coupons: Coupon[];
   selectedCoupon: Coupon | null;
   applyCoupon: (coupon: Coupon) => void; // ✅ 오타 수정
@@ -26,131 +31,99 @@ type CartPageProps = {
   };
 
   completeOrder: () => void;
+  addNotification: (message: string, type?: NotificationType) => void;
 };
 
 export default function CartPage({
   products,
   filteredProducts,
-  debouncedSearchTerm,
+  searchProductName,
   getRemainingStock,
   formatPrice,
-  addToCart,
   calculateItemTotal,
   cart,
-  removeFromCart,
-  updateQuantity,
   coupons,
   selectedCoupon,
-  totals,
-  applyCoupon,
   setSelectedCoupon,
   completeOrder,
+  setCart,
+  addNotification,
 }: CartPageProps) {
+  const calculateCartTotal = (): {
+    totalBeforeDiscount: number;
+    totalAfterDiscount: number;
+  } => {
+    let totalBeforeDiscount = 0;
+    let totalAfterDiscount = 0;
+
+    cart.forEach((item) => {
+      const itemPrice = item.product.price * item.quantity;
+      totalBeforeDiscount += itemPrice;
+      totalAfterDiscount += calculateItemTotal(item);
+    });
+
+    if (selectedCoupon) {
+      if (selectedCoupon.discountType === 'amount') {
+        totalAfterDiscount = Math.max(0, totalAfterDiscount - selectedCoupon.discountValue);
+      } else {
+        totalAfterDiscount = Math.round(
+          totalAfterDiscount * (1 - selectedCoupon.discountValue / 100)
+        );
+      }
+    }
+
+    return {
+      totalBeforeDiscount: Math.round(totalBeforeDiscount),
+      totalAfterDiscount: Math.round(totalAfterDiscount),
+    };
+  };
+
+  const totals = calculateCartTotal();
+
+  const addToCart = useCallback(
+    (product: ProductWithUI) => {
+      const remainingStock = getRemainingStock(product);
+      if (remainingStock <= 0) {
+        addNotification('재고가 부족합니다!', 'error');
+        return;
+      }
+
+      setCart((prevCart) => {
+        const existingItem = prevCart.find((item) => item.product.id === product.id);
+
+        if (existingItem) {
+          const newQuantity = existingItem.quantity + 1;
+
+          if (newQuantity > product.stock) {
+            addNotification(`재고는 ${product.stock}개까지만 있습니다.`, 'error');
+            return prevCart;
+          }
+
+          return prevCart.map((item) =>
+            item.product.id === product.id ? { ...item, quantity: newQuantity } : item
+          );
+        }
+
+        return [...prevCart, { product, quantity: 1 }];
+      });
+
+      addNotification('장바구니에 담았습니다', 'success');
+    },
+    [cart, addNotification, getRemainingStock]
+  );
+
   return (
     <div className='grid grid-cols-1 lg:grid-cols-4 gap-6'>
       <div className='lg:col-span-3'>
         {/* 상품 목록 */}
-        <section>
-          <div className='mb-6 flex justify-between items-center'>
-            <h2 className='text-2xl font-semibold text-gray-800'>전체 상품</h2>
-            <div className='text-sm text-gray-600'>총 {products.length}개 상품</div>
-          </div>
-          {filteredProducts.length === 0 ? (
-            <div className='text-center py-12'>
-              <p className='text-gray-500'>"{debouncedSearchTerm}"에 대한 검색 결과가 없습니다.</p>
-            </div>
-          ) : (
-            <div className='grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-4'>
-              {filteredProducts.map((product) => {
-                const remainingStock = getRemainingStock(product);
-
-                return (
-                  <div
-                    key={product.id}
-                    className='bg-white rounded-lg border border-gray-200 overflow-hidden hover:shadow-lg transition-shadow'
-                  >
-                    {/* 상품 이미지 영역 (placeholder) */}
-                    <div className='relative'>
-                      <div className='aspect-square bg-gray-100 flex items-center justify-center'>
-                        <svg
-                          className='w-24 h-24 text-gray-300'
-                          fill='none'
-                          stroke='currentColor'
-                          viewBox='0 0 24 24'
-                        >
-                          <path
-                            strokeLinecap='round'
-                            strokeLinejoin='round'
-                            strokeWidth={1}
-                            d='M4 16l4.586-4.586a2 2 0 012.828 0L16 16m-2-2l1.586-1.586a2 2 0 012.828 0L20 14m-6-6h.01M6 20h12a2 2 0 002-2V6a2 2 0 00-2-2H6a2 2 0 00-2 2v12a2 2 0 002 2z'
-                          />
-                        </svg>
-                      </div>
-                      {product.isRecommended && (
-                        <span className='absolute top-2 right-2 bg-red-500 text-white text-xs px-2 py-1 rounded'>
-                          BEST
-                        </span>
-                      )}
-                      {product.discounts.length > 0 && (
-                        <span className='absolute top-2 left-2 bg-orange-500 text-white text-xs px-2 py-1 rounded'>
-                          ~{Math.max(...product.discounts.map((d) => d.rate)) * 100}%
-                        </span>
-                      )}
-                    </div>
-
-                    {/* 상품 정보 */}
-                    <div className='p-4'>
-                      <h3 className='font-medium text-gray-900 mb-1'>{product.name}</h3>
-                      {product.description && (
-                        <p className='text-sm text-gray-500 mb-2 line-clamp-2'>
-                          {product.description}
-                        </p>
-                      )}
-
-                      {/* 가격 정보 */}
-                      <div className='mb-3'>
-                        <p className='text-lg font-bold text-gray-900'>
-                          {formatPrice(product.price, product.id)}
-                        </p>
-                        {product.discounts.length > 0 && (
-                          <p className='text-xs text-gray-500'>
-                            {product.discounts[0].quantity}개 이상 구매시 할인{' '}
-                            {product.discounts[0].rate * 100}%
-                          </p>
-                        )}
-                      </div>
-
-                      {/* 재고 상태 */}
-                      <div className='mb-3'>
-                        {remainingStock <= 5 && remainingStock > 0 && (
-                          <p className='text-xs text-red-600 font-medium'>
-                            품절임박! {remainingStock}개 남음
-                          </p>
-                        )}
-                        {remainingStock > 5 && (
-                          <p className='text-xs text-gray-500'>재고 {remainingStock}개</p>
-                        )}
-                      </div>
-
-                      {/* 장바구니 버튼 */}
-                      <button
-                        onClick={() => addToCart(product)}
-                        disabled={remainingStock <= 0}
-                        className={`w-full py-2 px-4 rounded-md font-medium transition-colors ${
-                          remainingStock <= 0
-                            ? 'bg-gray-100 text-gray-400 cursor-not-allowed'
-                            : 'bg-gray-900 text-white hover:bg-gray-800'
-                        }`}
-                      >
-                        {remainingStock <= 0 ? '품절' : '장바구니 담기'}
-                      </button>
-                    </div>
-                  </div>
-                );
-              })}
-            </div>
-          )}
-        </section>
+        <ProductList
+          filteredProducts={filteredProducts}
+          products={products}
+          searchProductName={searchProductName}
+          getRemainingStock={getRemainingStock}
+          addToCart={addToCart}
+          formatPrice={formatPrice}
+        />
       </div>
 
       <div className='lg:col-span-1'>
@@ -168,123 +141,28 @@ export default function CartPage({
               장바구니
             </h2>
             {cart.length === 0 ? (
-              <div className='text-center py-8'>
-                <svg
-                  className='w-16 h-16 text-gray-300 mx-auto mb-4'
-                  fill='none'
-                  stroke='currentColor'
-                  viewBox='0 0 24 24'
-                >
-                  <path
-                    strokeLinecap='round'
-                    strokeLinejoin='round'
-                    strokeWidth={1}
-                    d='M16 11V7a4 4 0 00-8 0v4M5 9h14l1 12H4L5 9z'
-                  />
-                </svg>
-                <p className='text-gray-500 text-sm'>장바구니가 비어있습니다</p>
-              </div>
+              <EmptyCart />
             ) : (
-              <div className='space-y-3'>
-                {cart.map((item) => {
-                  const itemTotal = calculateItemTotal(item);
-                  const originalPrice = item.product.price * item.quantity;
-                  const hasDiscount = itemTotal < originalPrice;
-                  const discountRate = hasDiscount
-                    ? Math.round((1 - itemTotal / originalPrice) * 100)
-                    : 0;
-
-                  return (
-                    <div key={item.product.id} className='border-b pb-3 last:border-b-0'>
-                      <div className='flex justify-between items-start mb-2'>
-                        <h4 className='text-sm font-medium text-gray-900 flex-1'>
-                          {item.product.name}
-                        </h4>
-                        <button
-                          onClick={() => removeFromCart(item.product.id)}
-                          className='text-gray-400 hover:text-red-500 ml-2'
-                        >
-                          <svg
-                            className='w-4 h-4'
-                            fill='none'
-                            stroke='currentColor'
-                            viewBox='0 0 24 24'
-                          >
-                            <path
-                              strokeLinecap='round'
-                              strokeLinejoin='round'
-                              strokeWidth={2}
-                              d='M6 18L18 6M6 6l12 12'
-                            />
-                          </svg>
-                        </button>
-                      </div>
-                      <div className='flex items-center justify-between'>
-                        <div className='flex items-center'>
-                          <button
-                            onClick={() => updateQuantity(item.product.id, item.quantity - 1)}
-                            className='w-6 h-6 rounded border border-gray-300 flex items-center justify-center hover:bg-gray-100'
-                          >
-                            <span className='text-xs'>−</span>
-                          </button>
-                          <span className='mx-3 text-sm font-medium w-8 text-center'>
-                            {item.quantity}
-                          </span>
-                          <button
-                            onClick={() => updateQuantity(item.product.id, item.quantity + 1)}
-                            className='w-6 h-6 rounded border border-gray-300 flex items-center justify-center hover:bg-gray-100'
-                          >
-                            <span className='text-xs'>+</span>
-                          </button>
-                        </div>
-                        <div className='text-right'>
-                          {hasDiscount && (
-                            <span className='text-xs text-red-500 font-medium block'>
-                              -{discountRate}%
-                            </span>
-                          )}
-                          <p className='text-sm font-medium text-gray-900'>
-                            {Math.round(itemTotal).toLocaleString()}원
-                          </p>
-                        </div>
-                      </div>
-                    </div>
-                  );
-                })}
-              </div>
+              <CartItemList
+                cart={cart}
+                calculateItemTotal={calculateItemTotal}
+                getRemainingStock={getRemainingStock}
+                addNotification={addNotification}
+                setCart={setCart}
+                products={products}
+              />
             )}
           </section>
 
           {cart.length > 0 && (
             <>
-              <section className='bg-white rounded-lg border border-gray-200 p-4'>
-                <div className='flex items-center justify-between mb-3'>
-                  <h3 className='text-sm font-semibold text-gray-700'>쿠폰 할인</h3>
-                  <button className='text-xs text-blue-600 hover:underline'>쿠폰 등록</button>
-                </div>
-                {coupons.length > 0 && (
-                  <select
-                    className='w-full text-sm border border-gray-300 rounded px-3 py-2 focus:outline-none focus:border-blue-500'
-                    value={selectedCoupon?.code || ''}
-                    onChange={(e) => {
-                      const coupon = coupons.find((c) => c.code === e.target.value);
-                      if (coupon) applyCoupon(coupon);
-                      else setSelectedCoupon(null);
-                    }}
-                  >
-                    <option value=''>쿠폰 선택</option>
-                    {coupons.map((coupon) => (
-                      <option key={coupon.code} value={coupon.code}>
-                        {coupon.name} (
-                        {coupon.discountType === 'amount'
-                          ? `${coupon.discountValue.toLocaleString()}원`
-                          : `${coupon.discountValue}%`}
-                        )
-                      </option>
-                    ))}
-                  </select>
-                )}
-              </section>
+              <CouponSelector
+                coupons={coupons}
+                selectedCoupon={selectedCoupon}
+                setSelectedCoupon={setSelectedCoupon}
+                calculateCartTotal={calculateCartTotal}
+                addNotification={addNotification}
+              />
 
               <section className='bg-white rounded-lg border border-gray-200 p-4'>
                 <h3 className='text-lg font-semibold mb-4'>결제 정보</h3>
